Bugfixes + preliminary SA

+++ mdpsa/src/neighborhood/move_mm_to_any.rs
@@ -43,7 +43,7 @@ impl NeighborhoodFunction for MoveMMToAny {
             }
         }
 
-        ((state.working_obj_val() - obj_prev) as f64, change_tokens)
+        ((state.working_obj_val() as isize - obj_prev as isize) as f64, change_tokens)
     }
 }
