+++ mdpsa/src/neighborhood/move_rm.rs
@@ -47,7 +47,7 @@ impl NeighborhoodFunction for MoveRM {
             }
         }
 
-        ((state.working_obj_val() - obj_prev) as f64, change_tokens)
+        ((state.working_obj_val() as isize - obj_prev as isize) as f64, change_tokens)
     }
 }
