+++ mdpsa/src/neighborhood/move_rm_to_any.rs
@@ -42,7 +42,7 @@ impl NeighborhoodFunction for MoveRMToAny {
             }
         }
 
-        ((state.working_obj_val() - obj_prev) as f64, change_tokens)
+        ((state.working_obj_val() as isize - obj_prev as isize) as f64, change_tokens)
     }
 }
